Move the function implementations for text and location content to eventcontent.cpp since the definitions are in eventcontent.h. This was missed when they were originally moved over.
